pyproject.toml
[project]
name = "mitre-technique-service"
version = "1.1.1"
description = "simplifies manipulation of OpenStack cloud platform for CyberRangeCZ Platform purposes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "cybersecurityhub.cz" }]
dependencies = [
    "djangorestframework",
    "taxii2-client",
    "structlog",
    "stix2",
    "yamlize",
    "ruamel.yaml<0.18",
    "redis",
    "jinja2",
    "django-cors-headers",
    "gunicorn",
    "drf-spectacular"
]

[project.urls]
GitHub = "https://github.com/cyberrangecz/backend-mitre-technique-service"

[build-system]
# TODO: Replace this with uv_build when it is it released for GA.
#   https://github.com/astral-sh/uv/issues/3957
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crczp"]
exclude-packages = [
    "crczp/mitre_matrix_visualizer_app/tests",
    "crczp/mitre_technique_project/tests",
]

[tool.ruff]
line-length = 100
preview = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
strict = true
pretty = true
enable_error_code = ["deprecated", "explicit-override"]
explicit_package_bases = true

[tool.pylint]
# You can add pylint configs here or in .pylintrc

[tool.bandit]
skips = ["B101"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "crczp/mitre_matrix_visualizer_app/tests",
    "crczp/mitre_technique_project/tests",
]
